+++ R/zzz.R
@@ -12,7 +12,7 @@
   make_decision_tree_rpart()
   make_decision_tree_partykit()
 
-  make_null_model_censored()
+  make_null_model_survival()
 
   make_proportional_hazards_survival()
   make_proportional_hazards_glmnet()